Add code for Carrot available on CWS: 2016-06-16
This is the code for the Chrome extension "Carrot", as contained in the
CRX file currently available from the Chrome Web Store.

This code is being uploaded to allow auditing of the code and is being
done in compliance with the license found in the `LICENSE` file (GPL
v3), which allows redistribution under the same terms.

This code is copyright (c) Five Industries 2016. I claim no copyright
over this code. The latest publicly available version of Carrot can be
found at the following address:

https://github.com/CarrotChatrooms/Carrot-Extension